+++ src/app/pages/donuts/donuts-id/donuts-id.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { AuthService } from 'src/app/service/auth.service';
 import { SweetService } from 'src/app/service/sweet.service';
 
 @Component({
@@ -11,11 +12,13 @@ export class DonutsIdComponent  implements OnInit {
   //trae e id y la info de donuts
   id!: number;
   donut!: any;
+  user: any;
 
-  constructor(private sweetService: SweetService, private activatedRoute: ActivatedRoute, private router: Router) { }
+  constructor(private sweetService: SweetService, private activatedRoute: ActivatedRoute, private router: Router,private authService: AuthService) { }
 
-
-  ngOnInit(): void {
+ngOnInit(): void {
+  // que usuario este conectado
+    this.user = this.authService.getUser()
     //mirar ruta y guardar id
     this.activatedRoute.paramMap.subscribe(params => {
       this.id = Number(params.get("id"));
@@ -28,8 +31,12 @@ export class DonutsIdComponent  implements OnInit {
   }
  //para eiminar el donuts
   delete(id: number) {
-    this.sweetService.deleteDonuts(id).subscribe(() => {
+    const confirmar = confirm("¿Estás seguro de que quieres eliminar?")
+    if(confirmar){
+      this.sweetService.deleteDonuts(id).subscribe(() => {
       this.router.navigate(["/"])
-    })
+      })
+    }
+   
   }
 }
